+++ src/axi_intf.sv
@@ -347,6 +347,7 @@ interface AXI_LITE #(
 
   // AW channel
   addr_t          aw_addr;
+  axi_pkg::prot_t aw_prot;
   logic           aw_valid;
   logic           aw_ready;
 
@@ -360,6 +361,7 @@ interface AXI_LITE #(
   logic           b_ready;
 
   addr_t          ar_addr;
+  axi_pkg::prot_t ar_prot;
   logic           ar_valid;
   logic           ar_ready;
 
@@ -369,18 +371,18 @@ interface AXI_LITE #(
   logic           r_ready;
 
   modport Master (
-    output aw_addr, aw_valid, input aw_ready,
+    output aw_addr, aw_prot, aw_valid, input aw_ready,
     output w_data, w_strb, w_valid, input w_ready,
     input b_resp, b_valid, output b_ready,
-    output ar_addr, ar_valid, input ar_ready,
+    output ar_addr, ar_prot, ar_valid, input ar_ready,
     input r_data, r_resp, r_valid, output r_ready
   );
 
   modport Slave (
-    input aw_addr, aw_valid, output aw_ready,
+    input aw_addr, aw_prot, aw_valid, output aw_ready,
     input w_data, w_strb, w_valid, output w_ready,
     output b_resp, b_valid, input b_ready,
-    input ar_addr, ar_valid, output ar_ready,
+    input ar_addr, ar_prot, ar_valid, output ar_ready,
     output r_data, r_resp, r_valid, input r_ready
   );
 
@@ -402,6 +404,7 @@ interface AXI_LITE_DV #(
 
   // AW channel
   addr_t          aw_addr;
+  axi_pkg::prot_t aw_prot;
   logic           aw_valid;
   logic           aw_ready;
 
@@ -415,6 +418,7 @@ interface AXI_LITE_DV #(
   logic           b_ready;
 
   addr_t          ar_addr;
+  axi_pkg::prot_t ar_prot;
   logic           ar_valid;
   logic           ar_ready;
 
@@ -424,18 +428,18 @@ interface AXI_LITE_DV #(
   logic           r_ready;
 
   modport Master (
-    output aw_addr, aw_valid, input aw_ready,
+    output aw_addr, aw_prot, aw_valid, input aw_ready,
     output w_data, w_strb, w_valid, input w_ready,
     input b_resp, b_valid, output b_ready,
-    output ar_addr, ar_valid, input ar_ready,
+    output ar_addr, ar_prot, ar_valid, input ar_ready,
     input r_data, r_resp, r_valid, output r_ready
   );
 
   modport Slave (
-    input aw_addr, aw_valid, output aw_ready,
+    input aw_addr, aw_prot, aw_valid, output aw_ready,
     input w_data, w_strb, w_valid, output w_ready,
     output b_resp, b_valid, input b_ready,
-    input ar_addr, ar_valid, output ar_ready,
+    input ar_addr, ar_prot, ar_valid, output ar_ready,
     output r_data, r_resp, r_valid, input r_ready
   );